changes to homepage template css
1) Top-picks:
Changed H1 in comic slider to figcaption, because the comic books are
figures, and the slider as already has an H1. Also made coresponding
changes to css.

2) Actiekaart:
Changed size of henks actiekaart H1 and 'order now' link

3) Ask Henk:
Changed position and class names of the henk-face, and ask henk logo.
finished speech bubbles.